# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title; generated import paths depend on it.
  title: Api
  version: 1.0.0
  description: PARKSYNC REST API contract. The Java Spring Boot service is the source of truth.
servers:
  - url: /api
    description: Replit development proxy and API context path
...
security:
  - bearerAuth: []
paths:
...
  /v1/profile:
    patch:
      operationId: updateProfile
      tags: [customer]
      summary: Update the signed-in user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /v1/profile/password:
    patch:
      operationId: changePassword
      tags: [customer]
      summary: Change the signed-in user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "204":
          description: Password changed
...
  /v1/vehicles:
    get:
      operationId: listVehicles
      tags: [customer]
      summary: List the signed-in user's vehicles
      responses:
        "200":
          description: Vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [customer]
      summary: Add a vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "201":
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /v1/vehicles/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    put:
      operationId: updateVehicle
      tags: [customer]
      summary: Update an owned vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [customer]
      summary: Delete an owned vehicle when it has no active booking
      responses:
        "204":
          description: Vehicle deleted
  /v1/locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: Search active parking facilities
      security: []
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: vehicleType
          in: query
          schema: { type: string }
        - name: availableOnly
          in: query
          schema: { type: boolean }
        - name: maxHourlyRate
          in: query
          schema: { type: number, minimum: 0 }
        - name: sort
          in: query
          schema: { type: string, enum: [price, availability, name] }
      responses:
        "200":
          description: Matching parking locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParkingLocation"
  /v1/locations/{id}:
    get:
      operationId: getLocation
      tags: [locations]
      summary: Get a parking location
      security: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingLocation"
  /v1/locations/{id}/slots:
    get:
      operationId: listLocationSlots
      tags: [locations]
      summary: List slot status and vehicle compatibility
      security: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Parking slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParkingSlot"
...
  /v1/payments/demo:
    post:
      operationId: createDemoPayment
      tags: [customer]
      summary: Record a clearly labelled demonstration payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoPaymentInput"
      responses:
        "201":
          description: Demo payment record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
  /v1/notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [customer]
      summary: Mark a notification as read
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Notification marked read
  /v1/support:
    get:
      operationId: listMySupportTickets
      tags: [customer]
      summary: List the signed-in user's support requests
      responses:
        "200":
          description: Support requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createSupportTicket
      tags: [customer]
      summary: Submit a support request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportInput"
      responses:
        "201":
          description: Support request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
...
  /v1/admin/locations/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    put:
      operationId: updateLocation
      tags: [admin]
      summary: Update a facility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "200":
          description: Updated facility
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingLocation"
    delete:
      operationId: deactivateLocation
      tags: [admin]
      summary: Deactivate a facility without deleting history
      responses:
        "204":
          description: Facility deactivated
...
  /v1/admin/locations/{id}/slots:
    get:
      operationId: listAdminLocationSlots
      tags: [admin]
      summary: List slots for an active or inactive facility
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Parking slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParkingSlot"
    post:
      operationId: createSlot
      tags: [admin]
      summary: Add a parking slot
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlotInput"
      responses:
        "201":
          description: Slot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingSlot"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, active, createdAt]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        role: { type: string, enum: [CUSTOMER, ADMIN] }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    ProfileUpdate:
      type: object
      required: [name, phone]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        phone: { type: string, minLength: 7, maxLength: 30 }
    PasswordChangeInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 10 }
    Vehicle:
      type: object
      required: [id, registration, type, brand, model]
      properties:
        id: { type: integer, format: int64 }
        registration: { type: string }
        type: { type: string, enum: [CAR, MOTORCYCLE, SCOOTER, OTHER] }
        brand: { type: string }
        model: { type: string }
        color: { type: ["string", "null"] }
    VehicleInput:
      type: object
      required: [registration, type, brand, model]
      properties:
        registration: { type: string, minLength: 4, maxLength: 20 }
        type: { type: string, enum: [CAR, MOTORCYCLE, SCOOTER, OTHER] }
        brand: { type: string, minLength: 1, maxLength: 80 }
        model: { type: string, minLength: 1, maxLength: 80 }
        color: { type: ["string", "null"] }
    ParkingLocation:
      type: object
      required: [id, name, address, category, description, operatingHours, hourlyRate, totalSlots, availableSlots, occupiedSlots, active, demoData]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        address: { type: string }
        area: { type: string }
        category: { type: string }
        description: { type: string }
        operatingHours: { type: string }
        supportedVehicleTypes:
          type: array
          items: { type: string }
        hourlyRate: { type: number, format: double }
        totalSlots: { type: integer }
        availableSlots: { type: integer }
        occupiedSlots: { type: integer }
        active: { type: boolean }
        demoData: { type: boolean }
        mapUrl: { type: ["string", "null"], format: uri }
    LocationInput:
      type: object
      required: [name, address, area, category, description, operatingHours, hourlyRate, supportedVehicleTypes]
      properties:
        name: { type: string, minLength: 2, maxLength: 150 }
        address: { type: string, minLength: 5, maxLength: 300 }
        area: { type: string }
        category: { type: string, enum: [MALL, PUBLIC, STANDALONE, OTHER] }
        description: { type: string, maxLength: 1000 }
        operatingHours: { type: string }
        supportedVehicleTypes:
          type: array
          items: { type: string, enum: [CAR, MOTORCYCLE, SCOOTER, OTHER] }
        hourlyRate: { type: number, minimum: 0.01 }
        mapUrl: { type: ["string", "null"], format: uri }
    ParkingSlot:
      type: object
      required: [id, locationId, code, status, vehicleTypes]
      properties:
        id: { type: integer, format: int64 }
        locationId: { type: integer, format: int64 }
        code: { type: string }
        status: { type: string, enum: [AVAILABLE, OCCUPIED, RESERVED, DISABLED] }
        vehicleTypes:
          type: array
          items: { type: string }
        hourlyRate: { type: ["number", "null"], format: double }
    SlotInput:
      type: object
      required: [code, status, vehicleTypes]
      properties:
        code: { type: string }
        status: { type: string, enum: [AVAILABLE, OCCUPIED, RESERVED, DISABLED] }
        vehicleTypes:
          type: array
          items: { type: string, enum: [CAR, MOTORCYCLE, SCOOTER, OTHER] }
        hourlyRate: { type: ["number", "null"], minimum: 0.01 }
...
    Payment:
      type: object
      required: [id, reference, bookingReference, amount, method, status, demo, createdAt]
      properties:
        id: { type: integer, format: int64 }
        reference: { type: string }
        bookingReference: { type: string }
        amount: { type: number, format: double }
        method: { type: string }
        status: { type: string }
        demo: { type: boolean }
        createdAt: { type: string, format: date-time }
    DemoPaymentInput:
      type: object
      required: [bookingId, method]
      properties:
        bookingId: { type: integer, format: int64 }
        method: { type: string, enum: [DEMO_CASH, DEMO_CARD, DEMO_UPI] }
...
    SupportInput:
      type: object
      required: [category, subject, description]
      properties:
        category: { type: string }
        subject: { type: string }
        description: { type: string, minLength: 10, maxLength: 3000 }
...
    SupportTicket:
      type: object
      required: [id, category, subject, description, status, createdAt]
      properties:
        id: { type: integer, format: int64 }
        customerName: { type: string }
        category: { type: string }
        subject: { type: string }
        description: { type: string }
        response: { type: ["string", "null"] }
        status: { type: string, enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED] }
        createdAt: { type: string, format: date-time }
        messages:
          type: array
          items:
            $ref: "#/components/schemas/SupportMessage"
    SupportMessage:
      type: object
      required: [id, authorName, authorRole, message, createdAt]
      properties:
        id: { type: integer, format: int64 }
        authorName: { type: string }
        authorRole: { type: string, enum: [CUSTOMER, ADMIN] }
        message: { type: string }
        createdAt: { type: string, format: date-time }
...